Created tests for  SearchBarComponent
